+++ src/Plugin.php
@@ -34,7 +34,7 @@ class Plugin
      *
      * @var string
      */
-    const VERSION = '3.3.1';
+    const VERSION = '3.3.2';
 
     /**
      * Constructor